Move tableToArray to utils/
